Add files via upload
I am creating basic login and sign in pages using the home pages, u can easily access that by using button in the home pages. if you click login in home pages it will redirected to that login page or if you click on signin page it will go to that signin page.